% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{EEG}
\alias{EEG}
\title{EEG Measurements in Patients with Alzheimer's Disease (long format)}
\format{
A data frame with 960 rows and 7 variables:
\describe{
  \item{resp}{EEG measurements}
  \item{sex}{sex of the patient}
  \item{age}{age of the patient, coded as 0 for less than 70 years and 1 for \eqn{>= 70} years}
  \item{diagnosis}{neuropsychological diagnosis, AD for Alzheimer's Disease, MCI for mild cognitive impairment or SCC for subjective cognitive complaints without clinically significant deficits}
  \item{region}{brain region of the EEG measurements, one of "temporal", "frontal" and "central"}
  \item{feature}{feature of the EEG measurements, either "brainrate" or "complexity"}
  \item{id}{Subject id}
}
}
\source{
Bathke, A., Friedrich, S., Konietschke, F., Pauly, M., Staffen, W., Strobl, N. and 
  Hoeller, Y. (2018). Testing Mean Differences among Groups: Multivariate and Repeated 
  Measures Analysis with Minimal Assumptions. Multivariate Behavioral Research. 
  Doi: 10.1080/00273171.2018.1446320.
}
\usage{
data(EEG)
}
\description{
At the Department of Neurology, University Clinic of Salzburg, 160 patients were diagnosed
with either AD, MCI, or SCC, based on neuropsychological diagnostics. This data set contains z-scores for brain rate and Hjorth complexity,
each measured at frontal, temporal and central electrode positions and averaged across hemispheres. In addition to standardization, complexity
values were multiplied by -1 in order to make them more easily comparable to brain rate
values: For brain rate we know that the values decrease with age and pathology, while
Hjorth complexity values are known to increase with age and pathology.
The three between-subjects factors considered were sex (men vs. women), diagnosis (AD
vs. MCI vs. SCC), and age (\eqn{< 70} vs. \eqn{>= 70} years). Additionally, the within-subjects factors region (frontal, temporal, central) and 
feature (brain rate, complexity) structure the response vector.
}
\examples{
if(requireNamespace("ggplot2")){ 
  library(ggplot2)
  ggplot(EEG, aes(x=sex, y=resp)) + geom_point(alpha=0.5) + facet_grid(region+feature~diagnosis) +
     stat_summary(fun.y = mean, fun.ymin = min, fun.ymax = max, colour = "red")
  }
  
}
\keyword{datasets}
