% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{plot.RM}
\alias{plot.RM}
\title{Plot function for an RM object}
\usage{
\method{plot}{RM}(x, leg = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{RM}}

\item{leg}{Logical: Should a legend be plotted?}

\item{...}{Additional parameters to be passed to plot()}
}
\description{
Generic plot function for \code{RM} objects: Returns a plot of the mean 
values along with confidence intervals for a specified RM-model.
}
