% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{print.MANOVA}
\alias{print.MANOVA}
\title{Display MANOVA object}
\usage{
\method{print}{MANOVA}(x, ...)
}
\arguments{
\item{x}{A MANOVA object}

\item{...}{Additional parameters (currently not used)}
}
\description{
Returns a short summary of the results (test statistics with p-values)
}
