% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{print.RM}
\alias{print.RM}
\title{Display an RM object}
\usage{
\method{print}{RM}(x, ...)
}
\arguments{
\item{x}{An RM object}

\item{...}{Additional parameters (currently not used)}
}
\description{
Returns a short summary of the results (test statistics with p-values)
}
