% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RM_MCTP.R
\name{RM.MCTP}
\alias{RM.MCTP}
\title{Post-hoc comparisons and simultaneous confidence intervals for 
contrasts in multivariate factorial repeated measure designs}
\usage{
RM.MCTP(
  object,
  alpha = 0.05,
  iter = 10000,
  contrast,
  contmat = NULL,
  type = NULL,
  factor_name = NULL,
  base = 1,
  seed = NULL
)
}
\arguments{
\item{object}{A \code{RM} object.}

\item{alpha}{Specifies the significance level, default is 0.05.}

\item{iter}{The number of iterations used for Wild Bootstrap. Default is 10000.}

\item{contrast}{The contrast matrix of interest. It can either be "pairwise" or "user-defined".}

\item{contmat}{If contrast = "user-defined", the contrast matrix must be specified here. Note that
its rows must sum to zero.}

\item{type}{If contrast is "pairwise", the type of the pairwise comparison must be specified here. 
Calculation is based on the contrMat function in package multcomp, see the corresponding help page 
for details on the types of contrasts available.}

\item{factor_name}{Specifies the factor for which post-hoc analysis are requested. Use All for all
pairwise comparisons.}

\item{base}{An integer specifying which group is considered the baseline group 
for Dunnett contrasts, see \code{\link[multcomp]{contrMat}}.}

\item{seed}{A random seed for the resampling procedure. Optional, for reproducibility of the bootstrap procedure.}
}
\value{
P-values and simultaneous confidence intervals for the chosen contrasts.
}
\description{
Post-hoc comparisons and simultaneous confidence intervals for 
contrasts in multivariate factorial repeated measure designs
}
\details{
The RM.MCTP() function computes p-values and confidence intervals for the chosen contrast with MCTP and with Wild Bootstrap.
}
\author{
Mona Modrzik
}
