% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{summary.MANOVA}
\alias{summary.MANOVA}
\title{Summarizing a MANOVA object}
\usage{
\method{summary}{MANOVA}(object, ...)
}
\arguments{
\item{object}{A MANOVA object}

\item{...}{Additional parameters (currently not used)}
}
\description{
Returns a summary of the results including mean vectors and sample sizes for all groups as well
as test statistics with degrees of freedom and p-values
}
