% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{summary.RM}
\alias{summary.RM}
\title{Summarizing an RM object}
\usage{
\method{summary}{RM}(object, ...)
}
\arguments{
\item{object}{An RM object}

\item{...}{Additional parameters (currently not used)}
}
\description{
Returns a summary of the results including mean values, variances and sample sizes for all groups as well
as test statistics with degrees of freedom and p-values
}
