% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{plot.RM}
\alias{plot.RM}
\title{Plot function for an RM object}
\usage{
\method{plot}{RM}(x, CI.info = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{RM}}

\item{CI.info}{If CI.info = TRUE, the mean values and confidence limits of the considered
contrast are printed.}

\item{...}{Additional parameters to be passed to plot()}
}
\description{
Generic plot function for \code{RM} objects: Returns a plot of the mean values along with confidence 
intervals for a factor (combination) specified by the user.
}
\details{
An additional argument \code{factor} can be used to specify the factor(s) used for plotting in two- and higher-way
layouts. See the examples for details.
}
